android-tools: 34.0.1 -> 34.0.4
The patch for linking with private `abseil-cpp` has been removed because
it is present in the upstream version already.

`ninja` has been added to the dependencies as building with `cmake`+`ninja`
is faster than with `cmake`+`make`.

The `-mod=vendor` go flag has been removed as the vendored version of
`boringssl` cannot be built as-is:

```
go: inconsistent vendoring in /build/android-tools-34.0.4/vendor/boringssl:
        golang.org/x/crypto@v0.6.0: is explicitly required in go.mod,
          but not marked as explicit in vendor/modules.txt
        golang.org/x/net@v0.7.0: is explicitly required in go.mod,
          but not marked as explicit in vendor/modules.txt
        golang.org/x/sys@v0.5.0: is explicitly required in go.mod,
          but not marked as explicit in vendor/modules.txt
        golang.org/x/term@v0.5.0: is explicitly required in go.mod,
          but not marked as explicit in vendor/modules.txt
```