stdenv: correctly determine SOURCE_DATE_EPOCH for more fetchers
Currently, `set-source-date-epoch-to-latest.sh` only correctly populates the
`SOURCE_DATE_EPOCH` environment variable for archives that are unpacked during
the derivation unpack phase. For fetchers that produce the source
already-unpacked, such as `fetchFromGitea` and `fetchFromGitHub`, the timestamp
is lost because the NAR format does not preserve file timestamps.

This change introduces a `keepSourceDate` option to those fetchers which
instructs them to retain the archive timestamp in
`out/nix-support/SOURCE_DATE_EPOCH`, and updates
`set-source-date-epoch-to-latest.sh` to pick it up.

To verify:
* does this work on darwin?
* does this work on other filesystem types?
* is this impacted by timezones anywhere?
* add tests for gitea(/forgejo/codeberg) fetcher and github fetcher
* check correct FOD behaviour
* check behaviour for derivations with multiple sources

I would like this to become the default and have a plan on how to achieve that
without breaking the FOD hashes, but that should probably be a follow-up PR
after trying this as a disabled-by-default feature for a while first.