Revert "Merge pull request #253876 from nbraud/nixos/sudo-rs"
This reverts commit 922926cfbc08f3e4065b51a41ebf613e59888015, reversing
changes made to 18c27868592bd0efd4aa131acd2bbda7034f94db.

As you can see on the PR discussion, this change sparked some
controversy, where the original maintainer asked for a full separation
of the `sudo` module and `sudo-rs` module, to which, the original PR
author refused to comply with.

Separations of NixOS modules in contexts of almost drop-in replacements
are not new in NixOS, prior art can be found in Forgejo/Gitea, see
248310 for an example.

Furthermore, they make sense when an "drop-in replacement" is not
totally a drop-in replacement, which is the case as this can be seen in
the issue tracker of `sudo-rs`.

In the meantime, until we obtain reasonable confidence that `sudo-rs` is
100% a drop-in replacement of `sudo`, it is not reasonable to push the
burden of a fused module maintenance on the original maintainer, even if
they did not commit anything to that module yet, which is not an
argument.

Moreover, as this can be seen in the discussion, the original maintainer
dropped the ball as they cannot reached a compromise with the original
PR author which said they would take care of this.

Nevertheless, it does not change that the PR was merged too fast (2
weeks) and caused cross compilation regressions because of ignored comments on
`buildPackages` handling which were raised early by the original
maintainer. Also, changes to the test driver were done without any look
or approval from the code owner of the test driver and without any idea
if the change can have bigger consequences.

The NixOS 23.11 release process will start in one month and two weeks,
this is too close to stabilize such a complicated change, no matter what
the original PR author believes it to be.

If such change has to pass while keeping the fused NixOS module, it will
have to wait after the 23.11 release so we have plenty of time of
testing it.

Otherwise, this change is free to be sent as another NixOS module as
recommended by the original maintainer and other NixOS developers in the
discussion.