go_1_21: fix cross-compiling other packages
Switch to distpack missed the fact that GOOS/GOARCH was set to
targetPlatform instead of hostPlatform so Go package used in
buildGoModule was built for the wrong platform (in this case,
build = host since we need to run Go on the build platform, but
host ≠ target when cross-compiling).

This change refactors build process for go1.21 to expose GOOS/GOARCH via
passhtru set to targetPlatform for compatiblity (since that’s what other
Go versions do), and build for hostPlatform as expected.

In addition to that, Go is a cross-compiler, that is, targetPlatform is
not relevant as it can build for any supported platform out-of-the-box,
and release builds are statically linked and do not require Cgo or C/C++
toolchain in general. Before this change, Go builds embedded C toolchain
reference for the targetPlatform (via {CC,CXX}_FOR_TARGET environment
variable). Now we expect that dependent packages configure CC/CXX for
their target platform (this should already be handled automatically when
using stdenv.mkDerivation and therefore buildGoModule).