openpgl: init at 0.5.0
Setting TBB_ROOT to the out path is necessary as otherwise the bundled
FindTBB.cmake tries to search the `tbb.dev` output for the library
files.

Note that embree and nanoflann are bundled within the source. The build
system is not currently setup to use non-vendored libs.

Rebuilt 25 times to test reproducibility.

Attempted a build on x86_64-darwin, and it failed, so marking this as
linux only.