hashlink: Add haxelib library to output
hashlink library is necessary to use the HL/C pipeline, allowing native compilation (in C with GCC) of haxe code.
The library is then detected by haxe, if in lib/haxe/{name}/{versionWithCommas}/...

Another way would be to do a separate package in haxePackages. but since it uses the same source I don't know if it's sane to not re-use.