diffoscope: fix broken build due to cbfs tests for now
The diffoscope build broke due to failing tests related to cbfs when
coreboot-utils was updated from 4.20 to 4.21 with commit
be6408699e598ea127c6a1a6da74ed27e95b819d.
Thanks to aij on github for reporting this issue and bisecting to find
the change in nixpkgs that introduced the issue.
The tests disabled in this commit should be re-enabled when diffoscope or
coreboot fix the underlying issue.

See: https://github.com/NixOS/nixpkgs/issues/256896