nixos/lib/test-driver: reduce spam at boot hangs
Since 008f9f0cd419bd66e922239e2319fd2b1f347ad8
 ("nixos/test-driver: actually use the backdoor message to wait for backdoor"),

when boot is still computering, we can get a tons of empty strings in response to the shell.

This is not really useful to print and waste the disk space for any CI system that logs them.

We stop logging chunks whenever they are empty.