fetch-npm-deps: Reduce flaky network issues by limiting parallelism
Sometimes fails flaky with errors like:

> [55] Failed sending data to the peer

that are difficult to reproduce at times, and cause a lot of
frustration as the general answer is to just keep retrying.

Testing with `nix-build --check -A bitwarden.npmDeps`, I got these
results:

- 2 cores: 20 / 20 runs succeeded in ~35 min total
- 4 cores: 18 / 20 runs succeeded in ~17 min total
- 8 cores: 11 / 20 runs succeeded in ~ 8 min total

This isn’t the ideal solution as it doesn’t fix the problem, but it is
at least something to temper the frequent flaky failures without
sacrificing too much runtime.

See
https://github.com/NixOS/nixpkgs/issues/256873
https://github.com/NixOS/nixpkgs/issues/257805