Fix version number in build.
cmake runs src/scmversion/gen_scmversion.sh which expects to be in a git repo, so scmversion.cpp is blank.
This leaves no build version in the title bar and help->about box.
This requires one to run the script in a cloned duckstation repo manually unless this can be automated