diffutils: disable tests on aarch64 musl
The tests have never worked on aarch64 musl, so this restores the
previous behaviour on that platform of building successfully even
though the tests don't pass.

Since the tests work on x86_64 musl, and not aarch64 musl, my best
guess having not looked too closely is that the difference is because
aarch64 is stuck on ancient bootstrap tools for musl that cause all
sorts of problems.  Hopefully we can revert this once the bootstrap
tools are updated.  Having diffutils building is a prerequisite to
doing that.

Fixes: 9588972fe983 ("diffutils: 3.9 -> 3.10")