bcachefs: fix version not showing correctly
buildLinux doesn't provide argsOverride by default.

Prior to 663caaa, a default kernel was replaced with a direct invocation
to buildLinux, this broke the use of argsOverride, in 663caaa, the use
of argsOverride was removed, and then later the invocation of buildLinux
was replaced with vanilla linux kernel that required argsOverride
which then rebroke things again.

Recommend either: put argsOverride inside buildLinux, or replace it with
something less confusing to maintainers.