Revert "Merge pull request #256497 from NickCao/python-msgpack"
This reverts commit 378fbaff985bb69424db700f0ec67aad8f83ad6f, reversing
changes made to 4b786de439279a1b09b2ba3256770d71b2698835.

As they caused breakage for borgbackup which is quite a ubiquitious
software for backupping and stopping all updates on the
nixos-unstable-small channel as it is hard to remove such software from
your configuration.

Thankfully, 1.0.6 and 1.0.7 changelog:

https://github.com/msgpack/msgpack-python/releases/tag/v1.0.6
https://github.com/msgpack/msgpack-python/releases/tag/v1.0.7

are not security updates, this should not cause issues.