python3Packages.msgpack: be mindful of borgbackup via passthru.tests
Unfortunately, BorgBackup, a backupping software is very sensible to MsgPack versions
to the extent of pinning patch-level versions…

I believe it's not totally absurd if they do it as they seems to consider it dangerous to
do otherwise.

As a result, I would like to ensure that python3Packages.msgpack are gated with that piece
of software ; of course, in case of security updates, we can consider extracting the patches
if reasonable or just breaking all dependents.