nsc: disable check phase on darwin
nsc is currently marked as broken on darwin because it's check phase
does not pass. However, it builds and works fine.

The check phase failure is due to a test in nsc stripping `-` from some
test output to remove some table formatting. This accidentally removes
`-` in a filesystem path and makes a test fail that expects the output
to contain a specific path.

This should be fixed upstream in nsc, but no need to mark darwin as
broken when it actually does build and work okay.