systemd: disable bootloader by default if EFI is disabled
If EFI is disabled, either because it is not available on the platform or
because the user has manually disabled it, automatically disable building
systemd-boot. The "withBootloader -> withEfi" assertion is maintained in
case someone manually tries to enable the bootloader without EFI support.

This fixes evaluation on platforms without EFI support.