glibc: use libgcc from pkgsHostHost
glibc depends on buildPackages.glibc for locale things.

buildPackages.glibc depended on buildPackages.libgcc, which, since
it's GCC, depends on the target's bintools, which depend on the
target's glibc, which, again, depends on buildPackages.glibc, causing
an infinute recursion when evaluating buildPackages.glibc when glibc
hasn't come from stdenv (e.g. on musl).

It's unlikely that buildPackages.glibc needs anything to do with the
target, so I've broken the cycle by having it use
pkgsHostHost.libgcc.

This fixes cross compilation from musl to glibc.