xvfb-run: only set fontconfig if not specified
Nix injects the environment variable FONTCONFIG_PATH to xvfb-run, in
order to provide some miminmal font support to applications via
fontconfig.

This change only sets this environment variable if it is not given
already. This allows the user to inject their own font configuration,
i.e. allows them to make extra fonts available to the executed
application.

If not set, the minimal font configuration will still be provided.