microsoft-edge: Change update script to have consistent ordering of versions
The current behaviour means it follows the order they are published in microsofts repos which sometimes changes on updates.
This changes it to sort by the branch name string which should be more consistent.

This patch doesn't change the ordering in default.nix will leave that for the next update patch to make that patch more easily backportable.