comictagger: init at 1.5.5
Adds comictagger, a multi-platform app for writing metadata to
digital comics, written in Python and PyQt.

The github project can be found here:
https://github.com/comictagger/comictagger

The following are python dependencies needed:
* py7zr
* pyinstaller
* pyinstaller-hooks-contrib
* settngs
* text2digits
* unrar-cffi
* wordninja

In this commit there are also those packages.

This is my first package, I expect "obvious" mistakes.

I've built and successfully run it on linux, for sure Mac dependencies
are different.

Note: To work on wayland it needs to be called from the command line
with a specific environemnt variable - I do not know how to handle it
at package level:

$ QT_QPA_PLATFORM=wayland comictagger