nixos/munin: fix key-context error when using extraAutoPlugins
If extraAutoPlugins contains values that carry context (e.g. it comes
from a flake input), the keys generated from them using baseNameOf
inherit that context and the config doesn't compile.

This doesn't actually need to be an attrset anyways, so a bit of
internal refactoring lets us fix this without changing the visible API.