systemd: allow building only the libraries
systemdMinimal is used to break the dependency cycle of libraries
required by systemd on itself. The problem is that this means that
every system closure will have two sets of systemd binaries in the
closure. This wastes around 10 MiB in every system closure.

Add the option to systemd to only ship libudev and libsystemd instead
of a whole systemd installation.