nixos/regreet: make `cage(1)` args configurable
In my case I'd like to be able to add `-m last` to `cage` to make sure
that the login form from regreet isn't displayed half on my external
monitor and half on my laptop screen, but on the last connected monitor
only.

That's basically the issue described in #226586, though it's not a
proper fix since the login form is shown on one monitor only.